backend/requirements.txt
Django>=3.0,<3.2
psycopg2>=2.8,<2.9
djangorestframework>=3.10.0,<3.12.4
django-environ
djangorestframework_simplejwt
djoser
pyyaml
django-cors-headers
graphene-django
PyJWT==1.7.0
django-graphql-jwt
elasticsearch-dsl>=6.0, <7.0
django-elasticsearch-dsl>=6.0,<7.0
graphene-elastic==0.7
django-filter
django-polymorphic
